notifier_test: stop checking backtrace content
Fixes #134 (Use fixtures for code hunk testing)

It turned out that we don't need fixtures to test that code. The expectations
can be simplified, so we just check that backtrace is not zero.

We already have fixtures that do what I had intended to
do (internal/testpkg1). We already cover code hunks in our tests.