Add ability to change reroute traffic at runtime
This is going to be used mostly by the remote configuration features. At the
moment you can configure `Host` and `APMHost` options to route your traffic,
however these values are static (cannot be changed at runtime).

With this commit we can redefine these two options and runtime and the traffic
will be redirected to the new location.