Fetch remote configuration every 10 minutes
Gobrake needs to be able to fetch remote configuration at certain
intervals. This commit prepares gobrake for the change. A newly created notifier
will start polling S3 for the config immediately. The config values are not used
at the moment, we just make reads and discard the config to keep the change as
simple as possible.