remote_config: delete the code for config dumping/loading
Fixes #185 (dumpConfig failed: no such file or directory)

Some of our users complain that gobrake tries to dump the remote config that we
pull from S3 to the GOPATH directory. The OS user that is running the code may
not have write access to that directory and therefore writing the remote config
will fail. This is especially evident when gobrake runs inside a
container. Persistency in a container makes no sense.

There's no good universal location to keep persistent remote config, so it's
easier just to delete the config.