Show proper error message which API gives 429 response
For the time being, if a response receives a 429 error, it will check to see if X-RateLimit-Delay is present or not in headers. If it is not, it will send the message 'X-RateLimit-Delay header is missing', otherwise it will compute the rate limit and send the static message 'IP is rate limited'.

Need to change
--------------

Replace the static message that says 'IP is rate limited' with the message that is received from the response body.